requirements.txt
jsonpath==0.82
PyYaml==5.3.1
configparser==5.0.0
requests==2.24.0
requests-toolbelt==0.9.1
pytest==6.0.1
PyMySQL==0.10.1
allure-pytest==2.8.19
cryptography==3.2.1
